% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app_server}
\alias{app_server}
\title{app_server
Server for Chemotaxis Dashboard}
\usage{
app_server(credentials, con)
}
\arguments{
\item{credentials}{data frame with credentials. Should have columns `user` and `password`}

\item{con}{database connection returned from `DBI::dbConnect`}
}
\value{
Server function for Chemotaxis Dashboard
}
\description{
app_server
Server for Chemotaxis Dashboard
}
