% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ces_tab.R
\name{ces_tab}
\alias{ces_tab}
\alias{ces_sidebarUI}
\alias{ces_cardsUI}
\alias{ces_server}
\title{ces_sidebarUI
UI element for the CE sidebar}
\usage{
ces_sidebarUI(id)

ces_cardsUI(id)

ces_server(
  id,
  con,
  shared_time_filter,
  shared_angle_filter,
  shared_track_len,
  shared_track_n,
  shared_ce_filter
)
}
\arguments{
\item{id}{Shiny namespace ID}

\item{con}{Active DBI database connection}

\item{shared_time_filter}{reactiveVal from the main server function for time filter definition}

\item{shared_angle_filter}{reactiveVal from the main server function for angle filter definition}

\item{shared_track_len}{reactiveVal from the main server function for physical track length filter in μm}

\item{shared_track_n}{reactiveVal from the main server function for track length filter for total number of frames}

\item{shared_ce_filter}{reactiveVal from the main server function for filtering on minimum chemotactic efficiency}
}
\value{
A modularized tagList

A modularized tagList of cards
}
\description{
ces_sidebarUI
UI element for the CE sidebar

ces_cardsUI
UI element for the CE cards

ces_server
Server logic for the Cross-Experiment Summary tab
}
