% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_quarto.R
\name{check_quarto}
\alias{check_quarto}
\title{check_quarto
Check that the desired version of quarto is installed}
\usage{
check_quarto(
  version = "1.5.57",
  version_warn = TRUE,
  version_error = !version_warn
)
}
\arguments{
\item{version}{character, desired version of quarto (use "" to check for any version of quarto)}

\item{version_warn}{logical, should a warning be issued if the installed version is less than the desired version? This is ignored if `version_error` is TRUE.}

\item{version_error}{logical, should an error be issued if the installed version is less than the desired version?}
}
\value{
TRUE if the desired version of quarto is installed, FALSE otherwise
}
\description{
check_quarto
Check that the desired version of quarto is installed
}
