% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{compare_prop_finish}
\alias{compare_prop_finish}
\title{compare_prop_finish
Compare the proportion of tracks that cross the lower ledge}
\usage{
compare_prop_finish(dat, comp_grp, trt = NULL, samp = NULL)
}
\arguments{
\item{dat}{data.frame of track data. Must contain the columns `treatment`, `sample`, `prop_finished`, `n_cells`, and the column to compare.}

\item{comp_grp}{character value of the column to use for comparison}

\item{trt}{character vector of the treatments to include in the comparison}

\item{samp}{character vector of the samples to include in the comparison}
}
\value{
A list of the fold change and p-value for each comparison
}
\description{
compare_prop_finish
Compare the proportion of tracks that cross the lower ledge
}
\details{
This function uses the prop.test function to compare the proportion of tracks that cross the lower ledge between two groups. If `trt` or `samp` are `NULL` all values in dat will be included.
}
