% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_experiments.R
\name{compare_two_functions}
\alias{compare_two_functions}
\title{Functional data analysis comparison of two functions, f1 and f2}
\usage{
compare_two_functions(
  data,
  frames.f,
  f,
  g,
  sig.figs,
  frames.g = frames.f,
  lab = NULL
)
}
\arguments{
\item{data}{A subset of dat_sub with variables, f and g, for permutation testing. When `lab` is specified, only `f` is used.}

\item{frames.f}{A vector of frames over which the function, f, is defined}

\item{f}{A vector or matrix of responses for f}

\item{g}{A vector or matrix of responses for g}

\item{sig.figs}{Integer, maximum significant digits for p-values obtained by permutation testing}

\item{frames.g}{A vector of frames over which the function, g, is defined}

\item{lab}{A character string indicating which variable is to be shuffled in the permutations. If NULL, then it is assumed that the two variables to be compared are `f` and `g`.}
}
\value{
A named vector containing the similarity measure and permutation p-value
.data <- filter(dat_sub, channel == 1) %>% dplyr::select(Track, Frame, x, y) %>% rename(f = x, g = y)
frames <- channel_summ$frames[[1]]
f <- channel_summ$x[[1]]
g <- channel_summ$y[[1]]
}
\description{
Functional data analysis comparison of two functions, f1 and f2
}
\details{
There are two main uses for this function: to compare two functions defined for each track (i.e. directed vs
undirected travel, where we shuffle un/directed values for each individual track) and to compare a single function 
between two different groups of tracks.

1) When comparing two functions defined for each track, we assume that there are two columns of data, labeled `f` 
and `g`. The will be compared, and in the permutation stage, they will be randomly swapped for about half of the 
Tracks in each permutation.

2) When comparing a single between two different groups of tracks, we assume that there is one column of data labeled,
`f`, and another column identified by `lab`. During the permutation stage, a group of tracks will be randomly assigned
group labels, maintaining the total number of Tracks in each group. Exactly two groups should be included in the data
set.
}
