% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{dbinit}
\alias{dbinit}
\title{dbinit
Set up database and infrastructure needed to run the Chemotaxis-Dashboard. If the database already exists and data is not empty, this will append the information in data to the database.}
\usage{
dbinit(db_path, data = NULL)
}
\arguments{
\item{db_path}{Character value specifying the path to the file where the database should be initialized.}

\item{data}{List of data.frames to initialize or add to the database. Internal test data will be used if `data` is NULL.}
}
\value{
Something - use this to sniff for existing DB information?
}
\description{
dbinit
Set up database and infrastructure needed to run the Chemotaxis-Dashboard. If the database already exists and data is not empty, this will append the information in data to the database.
}
\details{
The data.frames expected in data are `users`, `expSummary`, `expStats`, `chanSummary`, `chanRaw`,
`trackSummary`, and `trackRaw` (as returned by `process_experiments()`).
}
