% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{dbupdate}
\alias{dbupdate}
\title{dbupdate
Update the database with new data. This will append new information or update non key fields.}
\usage{
dbupdate(con, table, dat, key_fields)
}
\arguments{
\item{con}{DBIConnection object}

\item{table}{Character value specifying the table to update.}

\item{dat}{A data.frame to update or add to the database.}

\item{key_fields}{Character vector specifying the key fields in the table.}
}
\description{
dbupdate
Update the database with new data. This will append new information or update non key fields.
}
