% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_reports.R
\name{export_ses}
\alias{export_ses}
\title{Export reports based on data from a single experiment
export_ses}
\usage{
export_ses(
  experiment,
  db,
  format = "docx",
  file = experiment,
  file_ext = "docx",
  outdir = getwd(),
  seed = NULL,
  render = TRUE,
  export_csv = TRUE
)
}
\arguments{
\item{experiment}{character string, experiment ID}

\item{db}{character string, path to SQLite database}

\item{format}{character string, format for the report file (must be a format supported by quarto)}

\item{file}{character string, name for output files (not including the file extension)}

\item{file_ext}{character string, file extension for the report file}

\item{outdir}{character string, path to output directory}

\item{seed}{integer, random seed}

\item{render}{logical, whether to render the report. When false an unrendered qmd file is saved to the output directory.}

\item{export_csv}{logical, whether to export csv files for each channel in the experiment}
}
\description{
Export reports based on data from a single experiment
export_ses
}
\details{
This function exports reports based on data from a single experiment.
The report files are saved in the output directory and include csv files for each channel in the experiment and a report file of the specified file type.
The report file is generated using the `quarto` package and must be a format supported by `quarto` (see https://quarto.org/docs/reference/).
The location, `outdir`, must exist and the user must have write permissions to the directory.

If `render` is FALSE, then the report file is saved as an unrendered qmd file, and dependency checks for the rendering step will not be run.
}
