% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_thresh}
\alias{filter_thresh}
\title{filter_thresh}
\usage{
filter_thresh(trackRaw, thresh, invert = FALSE, drop = FALSE)
}
\arguments{
\item{trackRaw}{A data frame from the trackRaw table. Must contain the column `y`.}

\item{thresh}{A numeric value for the threshold.}

\item{invert}{A logical value. If FALSE (default), rows with y < thresh are flagged. If TRUE, rows with y > thresh are flagged.}

\item{drop}{A logical value. If FALSE (default), a new logical column `drop` is added to `trackRaw`. If TRUE, rows that are flagged are dropped.}
}
\value{
A data frame.
}
\description{
filter_thresh
}
