% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_dat}
\alias{get_dat}
\title{get_dat
Get data from the database, recpecting permissions}
\usage{
get_dat(con, user, select = "*", from, where = NULL)
}
\arguments{
\item{con}{DBI connection specifying the database to pull from}

\item{user}{Character value specifying the username}

\item{select}{Character value specifying the columns to pull}

\item{from}{Character value defining which table to pull from}

\item{where}{Character value that will get put into the WHERE part of the SQL statement}
}
\value{
A data.frame with the requested data
}
\description{
get_dat
Get data from the database, recpecting permissions
}
