% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{get_test_data}
\alias{get_test_data}
\title{get_test_data
Get a list of built-in test data from the package}
\usage{
get_test_data()
}
\value{
A list of test data as defined in `data-raw/process_raw_data.R`
}
\description{
get_test_data
Get a list of built-in test data from the package
}
