% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_experiment.R
\name{missing_info}
\alias{missing_info}
\title{function to display "Too much missing information" message in place of a regular figure}
\usage{
missing_info()
}
\value{
A ggplot2 figure
}
\description{
function to display "Too much missing information" message in place of a regular figure
}
