% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{observe_thresh}
\alias{observe_thresh}
\title{observe_thresh}
\usage{
observe_thresh(trackRaw, thresh)
}
\arguments{
\item{trackRaw}{A data frame from the trackRaw table. Must contain columns `expID`, `chanID`, `trackID`, and `y`.}

\item{thresh}{A numeric value for the threshold.}
}
\value{
The number of tracks that are observed crossing the threshold.
}
\description{
observe_thresh
}
