% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_experiment.R
\name{one_experiment}
\alias{one_experiment}
\title{Analysis of data for each experiment}
\usage{
one_experiment(dat_sub, root = "", sig.figs = 4)
}
\arguments{
\item{dat_sub}{data frame (tibble) containing the subset of data for a single experiment}

\item{root}{path to the root directory of this repository (or wherever else you want to be saving data - will save to `root/data/.`)}

\item{sig.figs}{maximum significant digits for p-values obtained by permutation testing}
}
\value{
A data frame containing channel-level summaries of the data in dat_sub. A list containing experiment-level statistics is also saved to an RData file.
}
\description{
Analysis of data for each experiment
}
