% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_nothing}
\alias{plot_nothing}
\title{plot_nothing
This will be called as a placeholder when there is nothing to plot}
\usage{
plot_nothing(message = "Please pick a subset to plot")
}
\arguments{
\item{message}{the message to display}
}
\value{
an empty ggplot object
}
\description{
plot_nothing
This will be called as a placeholder when there is nothing to plot
}
