% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_experiments.R
\name{process_uploaded_data}
\alias{process_uploaded_data}
\title{process_uploaded_data 
This will read in results from the convolutional model and preprocess for the dashboard (uploaded through the upload tab)}
\usage{
process_uploaded_data(
  uploaded_files,
  ledge_dist = 260,
  ledge_upper = 100,
  ledge_lower = 500
)
}
\arguments{
\item{uploaded_files}{A data.frame from the fileInput with the names and datapaths of the uploaded files.}

\item{ledge_dist}{Numeric, distance between top and bottom ledges of the microscope image in micrometers (default is 260)}

\item{ledge_upper}{Numeric, location of the upper ledge in the raw Y coordinate system (default is 100)}

\item{ledge_lower}{Numeric, location of the lower ledge in the raw Y coordinate system (default is 500)}
}
\value{
A list of data.frames...
}
\description{
process_uploaded_data 
This will read in results from the convolutional model and preprocess for the dashboard (uploaded through the upload tab)
}
