% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{qc_n_cells}
\alias{qc_n_cells}
\title{qc_n_cells
Figure to qc number of tracks/cells over time}
\usage{
qc_n_cells(dat)
}
\arguments{
\item{dat}{A data frame exported from the `trackRaw` table in the database}
}
\value{
A ggplot2 object
}
\description{
qc_n_cells
Figure to qc number of tracks/cells over time
}
