% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_tab.R
\name{qc_tab}
\alias{qc_tab}
\alias{qc_sidebarUI}
\alias{qc_cardsUI}
\alias{qc_server}
\title{qc_sidebarUI
UI element for the QC sidebar}
\usage{
qc_sidebarUI(id)

qc_cardsUI(id)

qc_server(
  id,
  con,
  shared_time_filter,
  shared_angle_filter,
  shared_track_len,
  shared_track_n,
  shared_ce_filter
)
}
\arguments{
\item{id}{Shiny namespace ID}

\item{con}{Active DBI database connection}

\item{shared_time_filter}{reactiveVal from the main server function for time filter definition}

\item{shared_angle_filter}{reactiveVal from the main server function for angle filter definition}

\item{shared_track_len}{reactiveVal from the main server function for physical track length definition}

\item{shared_track_n}{reactiveVal from the main server function for track length filter for total number of frames}

\item{shared_ce_filter}{reactiveVal from the main server function for filtering on minimum chemotactic efficiency}
}
\value{
A modularized tagList

A modularized tagList of cards
}
\description{
qc_sidebarUI
UI element for the QC sidebar

qc_cardsUI
UI element for the QC cards

qc_server
Server element for QC tab
}
