% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{qc_track_len}
\alias{qc_track_len}
\title{qc_track_len
Figure to qc track length}
\usage{
qc_track_len(dat)
}
\arguments{
\item{dat}{A data frame exported from the `trackRaw` table in the database}
}
\value{
A ggplot2 object
}
\description{
qc_track_len
Figure to qc track length
}
