% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{remove_dups}
\alias{remove_dups}
\title{remove_dups
Remove duplicate rows from the database}
\usage{
remove_dups(con, tabs = NULL)
}
\arguments{
\item{con}{A DBI database connection object}

\item{tabs}{A character vector of table names to remove duplicates from. If NULL, all tables will be checked.}
}
\description{
remove_dups
Remove duplicate rows from the database
}
