% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run a local instance of the Chemotaxis-Dashboard}
\usage{
run_app()
}
\description{
This function will start up a local instance of the Chemotaxis-Dashboard from within RStudio.
}
\details{
You can also use Docker (see https://github.com/IDSS-NIAID/JoesFlow for more documentation on using Docker), which includes a sample dataset.
}
