% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_plots.R
\name{ses_angle_migration}
\alias{ses_angle_migration}
\title{ses_angle_migration
Plot the net angle of migration}
\usage{
ses_angle_migration(dat)
}
\arguments{
\item{dat}{A data frame from the trackSummary table. Must contain columns `chanID`, `sID`, `treatment` and `angle_migration`.}
}
\value{
A ggplot object
}
\description{
ses_angle_migration
Plot the net angle of migration
}
