% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_plots.R
\name{ses_angle_migration_time}
\alias{ses_angle_migration_time}
\title{ses_angle_migration_time
Plot the angle of migration over time}
\usage{
ses_angle_migration_time(dat)
}
\arguments{
\item{dat}{A data frame from the trackRaw table. Must contain columns `chanID`, `sID`, `treatment`, `theta`, and `frames`.}
}
\value{
A ggplot object
}
\description{
ses_angle_migration_time
Plot the angle of migration over time
}
