% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_plots.R
\name{ses_chemotactic_efficiency}
\alias{ses_chemotactic_efficiency}
\title{ses_chemotactic_efficiency}
\usage{
ses_chemotactic_efficiency(dat)
}
\arguments{
\item{dat}{A data frame from the trackSummary table. Must contain columns `sID`, `chanID`, `trackID`, `treatment`, `ce`, and `frames`.}
}
\value{
A ggplot object
}
\description{
ses_chemotactic_efficiency
}
