% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_tab.R
\name{ses_tab}
\alias{ses_tab}
\alias{ses_sidebarUI}
\alias{ses_cardsUI}
\alias{ses_server}
\title{ses_sidebarUI
UI element for the SE sidebar}
\usage{
ses_sidebarUI(id)

ses_cardsUI(id)

ses_server(
  id,
  con,
  shared_time_filter,
  shared_angle_filter,
  shared_track_len,
  shared_track_n,
  shared_ce_filter
)
}
\arguments{
\item{id}{Shiny namespace ID}

\item{con}{Active DBI database connection}

\item{shared_time_filter}{reactiveVal from the main server function for time filter definition}

\item{shared_angle_filter}{reactiveVal from the main server function for angle filter definition}

\item{shared_track_len}{reactiveVal from the main server function for physical track length filter in μm}

\item{shared_track_n}{reactiveVal from the main server function for track length filter for total number of frames}

\item{shared_ce_filter}{reactiveVal from the main server function for filtering on minimum chemotactic efficiency}
}
\value{
A modularized tagList

A modularized tagList of cards
}
\description{
ses_sidebarUI
UI element for the SE sidebar

ses_cardsUI
UI element for the SE cards

ses_server
Server logic for the Single Experiment Statistics tab
}
