% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_plots.R
\name{ses_tracks_time}
\alias{ses_tracks_time}
\title{ses_tracks_time
Plot the number of tracks over time}
\usage{
ses_tracks_time(dat)
}
\arguments{
\item{dat}{A data frame from the trackRaw table. Must contain columns `sID`, `chanID`, `trackID`, `treatment`, `x`, `y`, and `frames`.}
}
\value{
A ggplot object
}
\description{
ses_tracks_time
Plot the number of tracks over time
}
