% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_plots.R
\name{ses_tracks_v}
\alias{ses_tracks_v}
\title{ses_tracks_v
Plot the velocity of tracks over time}
\usage{
ses_tracks_v(dat)
}
\arguments{
\item{dat}{A data frame from the trackRaw table. Must contain columns `sID`, `chanID`, `trackID`, `treatment`, `v_x`, `v_y`, and `frames`.}
}
\value{
A ggplot object
}
\description{
ses_tracks_v
Plot the velocity of tracks over time
}
