% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.management.R
\name{split_tracks}
\alias{split_tracks}
\title{Split Tracks}
\usage{
split_tracks(dat_sub, max_gap = 5)
}
\arguments{
\item{dat_sub}{A data frame containing raw tracking data.}

\item{max_gap}{The maximum number of frames that can be missing before a track is split. Default is 5 frames (2.5 minutes).}
}
\value{
A data frame with the same columns as `dat_sub`, but with tracks split
}
\description{
Identify and split tracks that have large periods of missing location data.
}
