% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{summarize_tracks}
\alias{summarize_tracks}
\title{summarize_tracks}
\usage{
summarize_tracks(trackRaw)
}
\arguments{
\item{trackRaw}{A data frame from the trackRaw table.}
}
\value{
A data frame with summarized track information.
}
\description{
summarize_tracks
}
