% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app_server}
\alias{app_server}
\title{app_server
Server for Chemotaxis Dashboard}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}
}
\value{
Server function for Chemotaxis Dashboard
}
\description{
app_server
Server for Chemotaxis Dashboard
}
