% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ces_tab.R
\name{ces_tab}
\alias{ces_tab}
\alias{ces_sidebarUI}
\alias{ces_cardsUI}
\alias{ces_server}
\title{ces_sidebarUI
UI element for the CE sidebar}
\usage{
ces_sidebarUI(id, con, user)

ces_cardsUI(id)

ces_server(id, con, user)
}
\arguments{
\item{id}{Shiny namespace ID}

\item{con}{Active DBI database connection}

\item{user}{Username of the user}
}
\value{
A modularized tagList

A modularized tagList of cards
}
\description{
ces_sidebarUI
UI element for the CE sidebar

ces_cardsUI
UI element for the CE cards

ces_server
Server logic for the Cross-Experiment Summary tab
}
