% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_experiments.R
\name{one_experiment}
\alias{one_experiment}
\title{Analysis of data for each experiment}
\usage{
one_experiment(
  dat_sub,
  experiment,
  results_dir,
  seed = NULL,
  sig.figs = 4,
  ledge_dist = 260
)
}
\arguments{
\item{dat_sub}{data frame (tibble) containing the subset of data for a single experiment}

\item{experiment}{Date of the experiment to run. Argument should be of the form: `experiment="\%Y\%m\%d"`.}

\item{results_dir}{path to directory for saving data}

\item{seed}{seed for random number generation}

\item{sig.figs}{maximum significant digits for p-values obtained by permutation testing}

\item{ledge_dist}{Numeric, distance between top and bottom ledge of the microscope image in micrometers (default is 260)}
}
\value{
A list of data.frames containing summaries and raw data from the data in dat_sub.
}
\description{
Analysis of data for each experiment
}
