% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_experiments.R
\name{one_experiment}
\alias{one_experiment}
\title{Analysis of data for each experiment}
\usage{
one_experiment(
  dat_sub,
  experiment,
  results_dir,
  ledge_dist = 260,
  ledge_upper = 100,
  ledge_lower = 500
)
}
\arguments{
\item{dat_sub}{data frame (tibble) containing the subset of data for a single experiment}

\item{experiment}{Date of the experiment to run. Argument should be of the form: `experiment="\%Y\%m\%d"`.}

\item{results_dir}{path to directory for saving data}

\item{ledge_dist}{Numeric, distance between top and bottom ledge of the microscope image in micrometers (default is 260)}

\item{ledge_upper}{Numeric, location of the upper ledge in the raw Y coordinate system (default is 100). This should be the first ledge crossed.}

\item{ledge_lower}{Numeric, location of the lower ledge in the raw Y coordinate system (default is 500). This should be the second ledge crossed.}
}
\value{
A list of data.frames containing summaries and raw data from the data in dat_sub.
}
\description{
Analysis of data for each experiment
}
