% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_tab.R
\name{qc_tab}
\alias{qc_tab}
\alias{qc_sidebarUI}
\alias{qc_cardsUI}
\alias{qc_server}
\title{qc_sidebarUI
UI element for the QC sidebar}
\usage{
qc_sidebarUI(id, con, user)

qc_cardsUI(id)

qc_server(id, con, user)
}
\arguments{
\item{id}{Shiny namespace ID}

\item{con}{Active DBI database connection}

\item{user}{Username of the user}
}
\value{
A modularized tagList

A modularized tagList of cards
}
\description{
qc_sidebarUI
UI element for the QC sidebar

qc_cardsUI
UI element for the QC cards

qc_server
Server element for QC tab
}
